3897 - Move new code to old page
