+++ src/client/pages/Admin/UserManagement/UserManagement.scss
@@ -50,3 +50,8 @@
     }
   }
 }
+
+.admin-user-management-cell {
+  text-align: left;
+  word-break: break-word;
+}
