--- src/client/pages/AdminUserManagement/AdminUserManagement.scss
@@ -1,7 +1,0 @@
-@import 'src/client/style/partials';
-
-.admin-user-management-cell {
-  text-align: left;
-  position: relative;
-  word-break: break-word;
-}
