--- src/client/pages/AdminUserManagement/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AdminUserManagement'
